+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Upcoming
+* Added max life config. Also added a new command argument to cap the maximum lives you can get
 * Added italian translation
 
 ## 1.0.2